+++ src/Indice.Features.Cases.UI/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Categories & Ordering for case types
+- WYSIWYG widget, using `ngx-quill`
 
 ## [6.11.3] - 2023-03-15
 ### Changed